Move upload functions to utils
