+++ .example.env
@@ -1,1 +1,3 @@
-AWS_S3_BUCKET=/*S3 BUCKET NAME*/
+AWS_ACCESS_KEY_ID="/***AWS ACCESS KEY ID***/"
+AWS_SECRET_ACCESS_KEY="/***AWS SECRET ACCESS KEY***/"
+AWS_S3_BUCKET="/***S3 BUCKET NAME**/"
